+++ rustler_codegen/src/init.rs
@@ -103,8 +103,14 @@ impl From<InitMacroInput> for proc_macro2::TokenStream {
                         load_info: rustler::codegen_runtime::NIF_TERM
                     ) -> rustler::codegen_runtime::c_int {
                         unsafe {
+                            let mut env = rustler::Env::new_init_env(&env, env);
                             // TODO: If an unwrap ever happens, we will unwind right into C! Fix this!
-                            rustler::codegen_runtime::handle_nif_init_call(#load, env, load_info)
+                            let load_info = rustler::Term::new(env, load_info);
+                            #load.map_or(0, |inner| {
+                                rustler::codegen_runtime::handle_nif_init_call(
+                                    inner, env, load_info
+                                )
+                            })
                         }
                     }
                     Some(nif_load)